address requested changes and improve interface responsivnes
