+++ packages/examples/packages/send-flow/LICENSE.APACHE2
@@ -186,7 +186,7 @@
       same "printed page" as the copyright notice for easier
       identification within third-party archives.
 
-   Copyright 2023 ConsenSys Software Inc.
+   Copyright 2024 Consensys Software Inc.
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
