+++ packages/examples/packages/send-flow/LICENSE.MIT0
@@ -1,6 +1,6 @@
 MIT No Attribution
 
-Copyright 2023 ConsenSys Software Inc.
+Copyright 2024 Consensys Software Inc.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this
 software and associated documentation files (the "Software"), to deal in the Software
