+++ packages/examples/packages/send-flow/src/components/SendFlowFooter.tsx
@@ -1,0 +1,12 @@
+import { Button, Footer, type SnapComponent } from '@metamask/snaps-sdk/jsx';
+
+/**
+ * A component that shows the send flow footer.
+ *
+ * @returns The SendFlowFooter component.
+ */
+export const SendFlowFooter: SnapComponent = () => (
+  <Footer>
+    <Button name="review">Review</Button>
+  </Footer>
+);
