+++ packages/examples/packages/send-flow/src/utils.tsx
@@ -37,7 +37,7 @@ export async function generateSendFlow({
           selectedCurrency="BTC"
           total={{ amount: 0, fiat: 0 }}
           fees={fees}
-          toAddress={null}
+          displayClearIcon={false}
         />
       ),
       context: {
